MIGENG-211 adding mocks to tests applying PR suggestions

+++ src/main/java/org/jboss/xavier/integrations/migrationanalytics/business/ParamsCalculator.java
@@ -25,7 +25,7 @@ public class ParamsCalculator implements Calculator<UploadFormInputDataModel> {
         Map mapa = (Map) e;
         Integer cputotalcores = (Integer) mapa.get(cpuTotalCoresPath);
         Integer cpucorespersocket = (Integer) mapa.get(cpuCoresPerSocketPath);
-        return Double.valueOf(Math.ceil(cputotalcores / (cpucorespersocket * 2.0))).intValue();
+        return (int) Math.ceil(cputotalcores / (cpucorespersocket * 2.0));
     }
 
     @Override
